requirements.txt
pandas
plotly
selenium
bs4
dash
dash_bootstrap_components
python-dotenv